refactor(AppController): Remove obsolete StateContext references; Relocate file to `component/`.
